refactor: dedupe code into modules for smaller vhs
