+++ lib/mp4/caption-parser.js
@@ -11,7 +11,10 @@
 
 var discardEmulationPreventionBytes = require('../tools/caption-packet-parser').discardEmulationPreventionBytes;
 var CaptionStream = require('../m2ts/caption-stream').CaptionStream;
-var inspect = require('../tools/mp4-inspector');
+var findBox = require('../mp4/find-box.js');
+var parseTfdt = require('../tools/parse-tfdt.js');
+var parseTrun = require('../tools/parse-trun.js');
+var parseTfhd = require('../tools/parse-tfhd.js');
 
 /**
   * Maps an offset in the mdat to a sample based on the the size of the samples.
@@ -133,7 +136,7 @@ var parseSamples = function(truns, baseMediaDecodeTime, tfhd) {
     // Note: We currently do not parse the sample table as well
     // as the trun. It's possible some sources will require this.
     // moov > trak > mdia > minf > stbl
-    var trackRun = inspect.parseTrun(trun);
+    var trackRun = parseTrun(trun);
     var samples = trackRun.samples;
 
     samples.forEach(function(sample) {
@@ -169,9 +172,9 @@ var parseSamples = function(truns, baseMediaDecodeTime, tfhd) {
  **/
 var parseCaptionNals = function(segment, videoTrackId) {
   // To get the samples
-  var trafs = inspect.findBox(segment, ['moof', 'traf']);
+  var trafs = findBox(segment, ['moof', 'traf']);
   // To get SEI NAL units
-  var mdats = inspect.findBox(segment, ['mdat']);
+  var mdats = findBox(segment, ['mdat']);
   var captionNals = {};
   var mdatTrafPairs = [];
 
@@ -187,14 +190,14 @@ var parseCaptionNals = function(segment, videoTrackId) {
   mdatTrafPairs.forEach(function(pair) {
     var mdat = pair.mdat;
     var traf = pair.traf;
-    var tfhd = inspect.findBox(traf, ['tfhd']);
+    var tfhd = findBox(traf, ['tfhd']);
     // Exactly 1 tfhd per traf
-    var headerInfo = inspect.parseTfhd(tfhd[0]);
+    var headerInfo = parseTfhd(tfhd[0]);
     var trackId = headerInfo.trackId;
-    var tfdt = inspect.findBox(traf, ['tfdt']);
+    var tfdt = findBox(traf, ['tfdt']);
     // Either 0 or 1 tfdt per traf
-    var baseMediaDecodeTime = (tfdt.length > 0) ? inspect.parseTfdt(tfdt[0]).baseMediaDecodeTime : 0;
-    var truns = inspect.findBox(traf, ['trun']);
+    var baseMediaDecodeTime = (tfdt.length > 0) ? parseTfdt(tfdt[0]).baseMediaDecodeTime : 0;
+    var truns = findBox(traf, ['trun']);
     var samples;
     var seiNals;
