+++ lib/mp4/probe.js
@@ -10,7 +10,11 @@
 
 var toUnsigned = require('../utils/bin').toUnsigned;
 var toHexString = require('../utils/bin').toHexString;
-var mp4Inspector = require('../tools/mp4-inspector.js');
+var findBox = require('../mp4/find-box.js');
+var parseType = require('../mp4/parse-type.js');
+var parseTfhd = require('../tools/parse-tfhd.js');
+var parseTrun = require('../tools/parse-trun.js');
+var parseTfdt = require('../tools/parse-tfdt.js');
 var timescale, startTime, compositionStartTime, getVideoTrackIds, getTracks,
   getTimescaleFromMediaHeader;
 
@@ -35,13 +39,13 @@ var timescale, startTime, compositionStartTime, getVideoTrackIds, getTracks,
 timescale = function(init) {
   var
     result = {},
-    traks = mp4Inspector.findBox(init, ['moov', 'trak']);
+    traks = findBox(init, ['moov', 'trak']);
 
   // mdhd timescale
   return traks.reduce(function(result, trak) {
     var tkhd, version, index, id, mdhd;
 
-    tkhd = mp4Inspector.findBox(trak, ['tkhd'])[0];
+    tkhd = findBox(trak, ['tkhd'])[0];
     if (!tkhd) {
       return null;
     }
@@ -52,7 +56,7 @@ timescale = function(init) {
                     tkhd[index + 2] <<  8 |
                     tkhd[index + 3]);
 
-    mdhd = mp4Inspector.findBox(trak, ['mdia', 'mdhd'])[0];
+    mdhd = findBox(trak, ['mdia', 'mdhd'])[0];
     if (!mdhd) {
       return null;
     }
@@ -86,11 +90,11 @@ startTime = function(timescale, fragment) {
   var trafs, baseTimes, result;
 
   // we need info from two childrend of each track fragment box
-  trafs = mp4Inspector.findBox(fragment, ['moof', 'traf']);
+  trafs = findBox(fragment, ['moof', 'traf']);
 
   // determine the start times for each track
   baseTimes = [].concat.apply([], trafs.map(function(traf) {
-    return mp4Inspector.findBox(traf, ['tfhd']).map(function(tfhd) {
+    return findBox(traf, ['tfhd']).map(function(tfhd) {
       var id, scale, baseTime;
 
       // get the track id from the tfhd
@@ -102,7 +106,7 @@ startTime = function(timescale, fragment) {
       scale = timescale[id] || 90e3;
 
       // get the base media decode time from the tfdt
-      baseTime = mp4Inspector.findBox(traf, ['tfdt']).map(function(tfdt) {
+      baseTime = findBox(traf, ['tfdt']).map(function(tfdt) {
         var version, result;
 
         version = tfdt[0];
@@ -151,7 +155,7 @@ startTime = function(timescale, fragment) {
  * @return {number} the composition start time for the fragment, in seconds
  **/
 compositionStartTime = function(timescales, fragment) {
-  var trafBoxes = mp4Inspector.findBox(fragment, ['moof', 'traf']);
+  var trafBoxes = findBox(fragment, ['moof', 'traf']);
   var baseMediaDecodeTime = 0;
   var compositionTimeOffset = 0;
   var trackId;
@@ -160,16 +164,18 @@ compositionStartTime = function(timescales, fragment) {
     // The spec states that track run samples contained within a `traf` box are contiguous, but
     // it does not explicitly state whether the `traf` boxes themselves are contiguous.
     // We will assume that they are, so we only need the first to calculate start time.
-    var parsedTraf = mp4Inspector.parseTraf(trafBoxes[0]);
+    var tfhd = parseTfhd(findBox(trafBoxes[0], ['tfhd'])[0]);
+    var trun = parseTrun(findBox(trafBoxes[0], ['trun'])[0]);
+    var tfdt = parseTfdt(findBox(trafBoxes[0], ['tfdt'])[0]);
 
-    for (var i = 0; i < parsedTraf.boxes.length; i++) {
-      if (parsedTraf.boxes[i].type === 'tfhd') {
-        trackId = parsedTraf.boxes[i].trackId;
-      } else if (parsedTraf.boxes[i].type === 'tfdt') {
-        baseMediaDecodeTime = parsedTraf.boxes[i].baseMediaDecodeTime;
-      } else if (parsedTraf.boxes[i].type === 'trun' && parsedTraf.boxes[i].samples.length) {
-        compositionTimeOffset = parsedTraf.boxes[i].samples[0].compositionTimeOffset || 0;
-      }
+    if (tfhd) {
+      trackId = tfhd.trackId;
+    }
+    if (tfdt) {
+      baseMediaDecodeTime = tfdt.baseMediaDecodeTime;
+    }
+    if (trun && trun.samples && trun.samples.length) {
+      compositionTimeOffset = trun.samples[0].compositionTimeOffset || 0;
     }
   }
 
@@ -193,15 +199,15 @@ compositionStartTime = function(timescales, fragment) {
   * @see ISO-BMFF-12/2015, Section 8.4.3
  **/
 getVideoTrackIds = function(init) {
-  var traks = mp4Inspector.findBox(init, ['moov', 'trak']);
+  var traks = findBox(init, ['moov', 'trak']);
   var videoTrackIds = [];
 
   traks.forEach(function(trak) {
-    var hdlrs = mp4Inspector.findBox(trak, ['mdia', 'hdlr']);
-    var tkhds = mp4Inspector.findBox(trak, ['tkhd']);
+    var hdlrs = findBox(trak, ['mdia', 'hdlr']);
+    var tkhds = findBox(trak, ['tkhd']);
 
     hdlrs.forEach(function(hdlr, index) {
-      var handlerType = mp4Inspector.parseType(hdlr.subarray(8, 12));
+      var handlerType = parseType(hdlr.subarray(8, 12));
       var tkhd = tkhds[index];
       var view;
       var version;
@@ -238,12 +244,12 @@ getTimescaleFromMediaHeader = function(mdhd) {
  * mp4 segment
  */
 getTracks = function(init) {
-  var traks = mp4Inspector.findBox(init, ['moov', 'trak']);
+  var traks = findBox(init, ['moov', 'trak']);
   var tracks = [];
 
   traks.forEach(function(trak) {
     var track = {};
-    var tkhd = mp4Inspector.findBox(trak, ['tkhd'])[0];
+    var tkhd = findBox(trak, ['tkhd'])[0];
     var view, tkhdVersion;
 
     // id
@@ -254,11 +260,11 @@ getTracks = function(init) {
       track.id = (tkhdVersion === 0) ? view.getUint32(12) : view.getUint32(20);
     }
 
-    var hdlr = mp4Inspector.findBox(trak, ['mdia', 'hdlr'])[0];
+    var hdlr = findBox(trak, ['mdia', 'hdlr'])[0];
 
     // type
     if (hdlr) {
-      var type = mp4Inspector.parseType(hdlr.subarray(8, 12));
+      var type = parseType(hdlr.subarray(8, 12));
 
       if (type === 'vide') {
         track.type = 'video';
@@ -271,14 +277,14 @@ getTracks = function(init) {
 
 
     // codec
-    var stsd = mp4Inspector.findBox(trak, ['mdia', 'minf', 'stbl', 'stsd'])[0];
+    var stsd = findBox(trak, ['mdia', 'minf', 'stbl', 'stsd'])[0];
 
     if (stsd) {
       var sampleDescriptions = stsd.subarray(8);
       // gives the codec type string
-      track.codec = mp4Inspector.parseType(sampleDescriptions.subarray(4, 8));
+      track.codec = parseType(sampleDescriptions.subarray(4, 8));
 
-      var codecBox = mp4Inspector.findBox(sampleDescriptions, [track.codec])[0];
+      var codecBox = findBox(sampleDescriptions, [track.codec])[0];
       var codecConfig, codecConfigType;
 
       if (codecBox) {
@@ -287,7 +293,7 @@ getTracks = function(init) {
           // we don't need anything but the "config" parameter of the
           // avc1 codecBox
           codecConfig = codecBox.subarray(78);
-          codecConfigType = mp4Inspector.parseType(codecConfig.subarray(4, 8));
+          codecConfigType = parseType(codecConfig.subarray(4, 8));
 
           if (codecConfigType === 'avcC' && codecConfig.length > 11) {
             track.codec += '.';
@@ -307,7 +313,7 @@ getTracks = function(init) {
         } else if ((/^mp4[a,v]$/i).test(track.codec)) {
           // we do not need anything but the streamDescriptor of the mp4a codecBox
           codecConfig = codecBox.subarray(28);
-          codecConfigType = mp4Inspector.parseType(codecConfig.subarray(4, 8));
+          codecConfigType = parseType(codecConfig.subarray(4, 8));
 
           if (codecConfigType === 'esds' && codecConfig.length > 20 && codecConfig[19] !== 0) {
             track.codec += '.' + toHexString(codecConfig[19]);
@@ -324,7 +330,7 @@ getTracks = function(init) {
       }
     }
 
-    var mdhd = mp4Inspector.findBox(trak, ['mdia', 'mdhd'])[0];
+    var mdhd = findBox(trak, ['mdia', 'mdhd'])[0];
 
     if (mdhd) {
       track.timescale = getTimescaleFromMediaHeader(mdhd);
@@ -338,8 +344,8 @@ getTracks = function(init) {
 
 module.exports = {
   // export mp4 inspector's findBox and parseType for backwards compatibility
-  findBox: mp4Inspector.findBox,
-  parseType: mp4Inspector.parseType,
+  findBox: findBox,
+  parseType: parseType,
   timescale: timescale,
   startTime: startTime,
   compositionStartTime: compositionStartTime,
