+++ lib/mp4/transmuxer.js
@@ -24,7 +24,6 @@ var isLikelyAacData = require('../aac/utils').isLikelyAacData;
 var ONE_SECOND_IN_TS = require('../utils/clock').ONE_SECOND_IN_TS;
 var AUDIO_PROPERTIES = require('../constants/audio-properties.js');
 var VIDEO_PROPERTIES = require('../constants/video-properties.js');
-var captionParser = require('./caption-parser.js');
 
 // object types
 var VideoSegmentStream, AudioSegmentStream, Transmuxer, CoalesceStream;
@@ -1179,6 +1178,5 @@ module.exports = {
   AUDIO_PROPERTIES: AUDIO_PROPERTIES,
   VIDEO_PROPERTIES: VIDEO_PROPERTIES,
   // exported for testing
-  generateVideoSegmentTimingInfo: generateVideoSegmentTimingInfo,
-  CaptionParser
+  generateVideoSegmentTimingInfo: generateVideoSegmentTimingInfo
 };
