+++ lib/tools/mp4-inspector.js
@@ -12,73 +12,11 @@
 var
   inspectMp4,
   textifyMp4,
-  toUnsigned = require('../utils/bin').toUnsigned,
   parseMp4Date = function(seconds) {
     return new Date(seconds * 1000 - 2082844800000);
   },
-  parseSampleFlags = function(flags) {
-    return {
-      isLeading: (flags[0] & 0x0c) >>> 2,
-      dependsOn: flags[0] & 0x03,
-      isDependedOn: (flags[1] & 0xc0) >>> 6,
-      hasRedundancy: (flags[1] & 0x30) >>> 4,
-      paddingValue: (flags[1] & 0x0e) >>> 1,
-      isNonSyncSample: flags[1] & 0x01,
-      degradationPriority: (flags[2] << 8) | flags[3]
-    };
-  },
-  /**
-   * Returns the string representation of an ASCII encoded four byte buffer.
-   * @param buffer {Uint8Array} a four-byte buffer to translate
-   * @return {string} the corresponding string
-   */
-  parseType = function(buffer) {
-    var result = '';
-    result += String.fromCharCode(buffer[0]);
-    result += String.fromCharCode(buffer[1]);
-    result += String.fromCharCode(buffer[2]);
-    result += String.fromCharCode(buffer[3]);
-    return result;
-  },
-  // Find the data for a box specified by its path
-  findBox = function(data, path) {
-    var results = [],
-        i, size, type, end, subresults;
-
-    if (!path.length) {
-      // short-circuit the search for empty paths
-      return null;
-    }
-
-    for (i = 0; i < data.byteLength;) {
-      size  = toUnsigned(data[i]     << 24 |
-                         data[i + 1] << 16 |
-                         data[i + 2] <<  8 |
-                         data[i + 3]);
-
-      type = parseType(data.subarray(i + 4, i + 8));
-
-      end = size > 1 ? i + size : data.byteLength;
-
-      if (type === path[0]) {
-        if (path.length === 1) {
-          // this is the end of the path and we've found the box we were
-          // looking for
-          results.push(data.subarray(i + 8, end));
-        } else {
-          // recursively search for the next box along the path
-          subresults = findBox(data.subarray(i + 8, end), path.slice(1));
-          if (subresults.length) {
-            results = results.concat(subresults);
-          }
-        }
-      }
-      i = end;
-    }
-
-    // we've finished searching all of data
-    return results;
-  },
+  parseType = require('../mp4/parse-type'),
+  findBox = require('../mp4/find-box'),
   nalParse = function(avcStream) {
     var
       avcView = new DataView(avcStream.buffer, avcStream.byteOffset, avcStream.byteLength),
@@ -438,33 +376,7 @@ var
       }
       return result;
     },
-    sidx: function(data) {
-      var view = new DataView(data.buffer, data.byteOffset, data.byteLength),
-          result = {
-            version: data[0],
-            flags: new Uint8Array(data.subarray(1, 4)),
-            references: [],
-            referenceId: view.getUint32(4),
-            timescale: view.getUint32(8),
-            earliestPresentationTime: view.getUint32(12),
-            firstOffset: view.getUint32(16)
-          },
-          referenceCount = view.getUint16(22),
-          i;
-
-      for (i = 24; referenceCount; i += 12, referenceCount--) {
-        result.references.push({
-          referenceType: (data[i] & 0x80) >>> 7,
-          referencedSize: view.getUint32(i) & 0x7FFFFFFF,
-          subsegmentDuration: view.getUint32(i + 4),
-          startsWithSap: !!(data[i + 8] & 0x80),
-          sapType: (data[i + 8] & 0x70) >>> 4,
-          sapDeltaTime: view.getUint32(i + 8) & 0x0FFFFFFF
-        });
-      }
-
-      return result;
-    },
+    sidx: require('./parse-sidx.js'),
     smhd: function(data) {
       return {
         version: data[0],
@@ -555,65 +467,8 @@ var
     styp: function(data) {
       return parse.ftyp(data);
     },
-    tfdt: function(data) {
-      var result = {
-        version: data[0],
-        flags: new Uint8Array(data.subarray(1, 4)),
-        baseMediaDecodeTime: toUnsigned(data[4] << 24 | data[5] << 16 | data[6] << 8 | data[7])
-      };
-      if (result.version === 1) {
-        result.baseMediaDecodeTime *= Math.pow(2, 32);
-        result.baseMediaDecodeTime += toUnsigned(data[8] << 24 | data[9] << 16 | data[10] << 8 | data[11]);
-      }
-      return result;
-    },
-    tfhd: function(data) {
-      var
-        view = new DataView(data.buffer, data.byteOffset, data.byteLength),
-        result = {
-          version: data[0],
-          flags: new Uint8Array(data.subarray(1, 4)),
-          trackId: view.getUint32(4)
-        },
-        baseDataOffsetPresent = result.flags[2] & 0x01,
-        sampleDescriptionIndexPresent = result.flags[2] & 0x02,
-        defaultSampleDurationPresent = result.flags[2] & 0x08,
-        defaultSampleSizePresent = result.flags[2] & 0x10,
-        defaultSampleFlagsPresent = result.flags[2] & 0x20,
-        durationIsEmpty = result.flags[0] & 0x010000,
-        defaultBaseIsMoof =  result.flags[0] & 0x020000,
-        i;
-
-      i = 8;
-      if (baseDataOffsetPresent) {
-        i += 4; // truncate top 4 bytes
-        // FIXME: should we read the full 64 bits?
-        result.baseDataOffset = view.getUint32(12);
-        i += 4;
-      }
-      if (sampleDescriptionIndexPresent) {
-        result.sampleDescriptionIndex = view.getUint32(i);
-        i += 4;
-      }
-      if (defaultSampleDurationPresent) {
-        result.defaultSampleDuration = view.getUint32(i);
-        i += 4;
-      }
-      if (defaultSampleSizePresent) {
-        result.defaultSampleSize = view.getUint32(i);
-        i += 4;
-      }
-      if (defaultSampleFlagsPresent) {
-        result.defaultSampleFlags = view.getUint32(i);
-      }
-      if (durationIsEmpty) {
-        result.durationIsEmpty = true;
-      }
-      if (!baseDataOffsetPresent && defaultBaseIsMoof) {
-        result.baseDataOffsetIsMoof = true;
-      }
-      return result;
-    },
+    tfdt: require('./parse-tfdt.js'),
+    tfhd: require('./parse-tfhd.js'),
     tkhd: function(data) {
       var
         view = new DataView(data.buffer, data.byteOffset, data.byteLength),
@@ -686,84 +541,7 @@ var
         sampleDegradationPriority: view.getUint16(22)
       };
     },
-    trun: function(data) {
-      var
-        result = {
-          version: data[0],
-          flags: new Uint8Array(data.subarray(1, 4)),
-          samples: []
-        },
-        view = new DataView(data.buffer, data.byteOffset, data.byteLength),
-        // Flag interpretation
-        dataOffsetPresent = result.flags[2] & 0x01, // compare with 2nd byte of 0x1
-        firstSampleFlagsPresent = result.flags[2] & 0x04, // compare with 2nd byte of 0x4
-        sampleDurationPresent = result.flags[1] & 0x01, // compare with 2nd byte of 0x100
-        sampleSizePresent = result.flags[1] & 0x02, // compare with 2nd byte of 0x200
-        sampleFlagsPresent = result.flags[1] & 0x04, // compare with 2nd byte of 0x400
-        sampleCompositionTimeOffsetPresent = result.flags[1] & 0x08, // compare with 2nd byte of 0x800
-        sampleCount = view.getUint32(4),
-        offset = 8,
-        sample;
-
-      if (dataOffsetPresent) {
-        // 32 bit signed integer
-        result.dataOffset = view.getInt32(offset);
-        offset += 4;
-      }
-
-      // Overrides the flags for the first sample only. The order of
-      // optional values will be: duration, size, compositionTimeOffset
-      if (firstSampleFlagsPresent && sampleCount) {
-        sample = {
-          flags: parseSampleFlags(data.subarray(offset, offset + 4))
-        };
-        offset += 4;
-        if (sampleDurationPresent) {
-          sample.duration = view.getUint32(offset);
-          offset += 4;
-        }
-        if (sampleSizePresent) {
-          sample.size = view.getUint32(offset);
-          offset += 4;
-        }
-        if (sampleCompositionTimeOffsetPresent) {
-          if (result.version === 1) {
-            sample.compositionTimeOffset = view.getInt32(offset);
-          } else {
-            sample.compositionTimeOffset = view.getUint32(offset);
-          }
-          offset += 4;
-        }
-        result.samples.push(sample);
-        sampleCount--;
-      }
-
-      while (sampleCount--) {
-        sample = {};
-        if (sampleDurationPresent) {
-          sample.duration = view.getUint32(offset);
-          offset += 4;
-        }
-        if (sampleSizePresent) {
-          sample.size = view.getUint32(offset);
-          offset += 4;
-        }
-        if (sampleFlagsPresent) {
-          sample.flags = parseSampleFlags(data.subarray(offset, offset + 4));
-          offset += 4;
-        }
-        if (sampleCompositionTimeOffsetPresent) {
-          if (result.version === 1) {
-            sample.compositionTimeOffset = view.getInt32(offset);
-          } else {
-            sample.compositionTimeOffset = view.getUint32(offset);
-          }
-          offset += 4;
-        }
-        result.samples.push(sample);
-      }
-      return result;
-    },
+    trun: require('./parse-trun.js'),
     'url ': function(data) {
       return {
         version: data[0],
